requirements.txt
boto3
diskcache
numpy
certifi
idna
fusepy
tenacity
